Move cleanser into single project
Update test files to accommodate and tidy
