--- event-processing/cleanser/.npmignore
@@ -1,1 +1,0 @@
-tests/*
