--- event-processing/cleanser/tsconfig.json
@@ -1,15 +1,0 @@
-{
-    "compilerOptions": {
-      "target": "es2020",
-      "strict": true,
-      "preserveConstEnums": true,
-      "noEmit": true,
-      "sourceMap": true,
-      "module":"es2015",
-      "moduleResolution":"node",
-      "esModuleInterop": true, 
-      "skipLibCheck": true,
-      "forceConsistentCasingInFileNames": true,  
-    },
-    "exclude": ["node_modules", "**/*.test.ts"]
-  }
