+++ event-processing/lambda/app.ts
@@ -15,6 +15,9 @@ import { EnrichmentService } from './services/enrichment-service';
 import { AuditEvent, IAuditEvent } from './models/audit-event';
 import { KeyService } from './services/key-service';
 import { ObfuscationService } from './services/obfuscation-service';
+import { ICleansedEvent } from './models/cleansed-event';
+import { EnrichedAuditEvent, IEnrichedAuditEvent } from './models/enriched-audit-event';
+import { CleansingService } from './services/cleansing-service';
 
 export const eventProcessorHandler = async (event: SQSEvent): Promise<void> => {
     for (const record of event.Records) {
@@ -89,6 +92,38 @@ export const obfuscationHandler = async (event: FirehoseTransformationEvent): Pr
     } else {
         console.log(`Processing completed. Successful records ${output.length}.`);
     }
+
+    return { records: output };
+};
+
+export const cleanserHandler = async (event: FirehoseTransformationEvent): Promise<FirehoseTransformationResult> => {
+    /* Process the list of records and transform them */
+    const transformationResult: FirehoseRecordTransformationStatus = 'Ok';
+
+    const output = event.records.map((record: FirehoseTransformationEventRecord) => {
+        const plaintextData: string = Buffer.from(record.data, 'base64').toString('ascii');
+        const events: unknown[] = JSON.parse(plaintextData);
+        const cleansedEvents: ICleansedEvent[] = [];
+        let data: string;
+
+        if (events.length > 0) {
+            for (const k in events) {
+                const auditEvent: IEnrichedAuditEvent = EnrichedAuditEvent.fromJSONString(JSON.stringify(events[k]));
+                cleansedEvents.push(CleansingService.cleanseEvent(auditEvent));
+            }
+            data = Buffer.from(JSON.stringify(cleansedEvents)).toString('base64');
+        } else {
+            const auditEvent: IEnrichedAuditEvent = EnrichedAuditEvent.fromJSONString(plaintextData);
+            const cleansedEvent = CleansingService.cleanseEvent(auditEvent);
+            data = Buffer.from(JSON.stringify(cleansedEvent)).toString('base64');
+        }
+
+        return {
+            recordId: record.recordId,
+            result: transformationResult,
+            data: data,
+        } as FirehoseTransformationResultRecord;
+    });
 
     return { records: output };
 };
