+++ event-processing/lambda/tests/test-helpers/cleanser-helper.ts
@@ -1,9 +1,10 @@
 /* istanbul ignore file */
-import { IAuditEvent } from '../../models/audit-event';
-import { ObfuscationService } from '../../services/obfuscation-service';
 
-export class ObfuscationHelper {
-    static exampleObfuscatedMessage: IAuditEvent = {
+import { IEnrichedAuditEvent } from '../../models/enriched-audit-event';
+import { ICleansedEvent } from '../../models/cleansed-event';
+
+export class CleanserHelper {
+    static exampleEnrichedMessage: IEnrichedAuditEvent = {
         event_id: '66258f3e-82fc-4f61-9ba0-62424e1f06b4',
         client_id: 'some-client',
         timestamp: 1609462861,
@@ -11,10 +12,9 @@ export class ObfuscationHelper {
         event_name: 'AUTHENTICATION_ATTEMPT',
         component_id: 'AUTH',
         user: {
-            transaction_id: ObfuscationService.obfuscateField('a52f6f87', 'secret-1-value'),
-            user_id: 'some_user_id',
-            email: ObfuscationService.obfuscateField('foo@bar.com', 'secret-1-value'),
-            phone: ObfuscationService.obfuscateField('07711223344', 'secret-1-value'),
+            transaction_id: 'a52f6f87',
+            email: 'foo@bar.com',
+            phone: '07711223344',
             ip_address: '100.100.100.100',
             session_id: 'c222c1ec',
             persistent_session_id: 'some session id',
@@ -24,10 +24,19 @@ export class ObfuscationHelper {
             xray_trace_id: '24727sda4192',
         },
         restricted: {
-            experian_ref: ObfuscationService.obfuscateField('DSJJSEE29392', 'secret-1-value'),
+            experian_ref: 'DSJJSEE29392',
+            passport_number: 1040349934,
         },
         extensions: {
             response: 'Authentication successful',
         },
+    };
+
+    static exampleCleansedMessage: ICleansedEvent = {
+        event_id: '66258f3e-82fc-4f61-9ba0-62424e1f06b4',
+        event_name: 'AUTHENTICATION_ATTEMPT',
+        component_id: 'AUTH',
+        timestamp: 1609462861,
+        timestamp_formatted: '2021-01-23T15:43:21.842',
     };
 }
