+++ event-processing/lambda/tests/test-helpers/event-processor-helper.ts
@@ -1,9 +1,9 @@
 /* istanbul ignore file */
-import { IAuditEvent } from '../../models/audit-event';
-import { ObfuscationService } from '../../services/obfuscation-service';
 
-export class ObfuscationHelper {
-    static exampleObfuscatedMessage: IAuditEvent = {
+import { IAuditEvent } from '../../models/audit-event';
+
+export class EventProcessorHelper {
+    static exampleAuditMessage: IAuditEvent = {
         event_id: '66258f3e-82fc-4f61-9ba0-62424e1f06b4',
         client_id: 'some-client',
         timestamp: 1609462861,
@@ -11,10 +11,10 @@ export class ObfuscationHelper {
         event_name: 'AUTHENTICATION_ATTEMPT',
         component_id: 'AUTH',
         user: {
-            transaction_id: ObfuscationService.obfuscateField('a52f6f87', 'secret-1-value'),
+            transaction_id: 'a52f6f87',
             user_id: 'some_user_id',
-            email: ObfuscationService.obfuscateField('foo@bar.com', 'secret-1-value'),
-            phone: ObfuscationService.obfuscateField('07711223344', 'secret-1-value'),
+            email: 'foo@bar.com',
+            phone: '07711223344',
             ip_address: '100.100.100.100',
             session_id: 'c222c1ec',
             persistent_session_id: 'some session id',
@@ -24,7 +24,7 @@ export class ObfuscationHelper {
             xray_trace_id: '24727sda4192',
         },
         restricted: {
-            experian_ref: ObfuscationService.obfuscateField('DSJJSEE29392', 'secret-1-value'),
+            experian_ref: 'DSJJSEE29392',
         },
         extensions: {
             response: 'Authentication successful',
