+++ event-processing/lambda/tests/test-helpers/test-helper.ts
@@ -40,35 +40,6 @@ export class TestHelper {
         },
     };
 
-    static exampleMessage: IAuditEvent = {
-        event_id: '66258f3e-82fc-4f61-9ba0-62424e1f06b4',
-        client_id: 'some-client',
-        timestamp: 1609462861,
-        timestamp_formatted: '2021-01-23T15:43:21.842',
-        event_name: 'AUTHENTICATION_ATTEMPT',
-        component_id: 'AUTH',
-        user: {
-            transaction_id: 'a52f6f87',
-            user_id: 'some_user_id',
-            email: 'foo@bar.com',
-            phone: '07711223344',
-            ip_address: '100.100.100.100',
-            session_id: 'c222c1ec',
-            persistent_session_id: 'some session id',
-            govuk_signin_journey_id: '43143-233Ds-2823-283-dj299j1',
-        },
-        platform: {
-            xray_trace_id: '24727sda4192',
-        },
-        restricted: {
-            experian_ref: 'DSJJSEE29392',
-            passport_number: 1040349934,
-        },
-        extensions: {
-            response: 'Authentication successful',
-        },
-    };
-
     private static firehoseTransformationEvent: FirehoseTransformationEvent = {
         invocationId: 'bb5d7530-f7b7-44aa-bc0f-01b76248fbb8',
         deliveryStreamArn: 'arn:aws:firehose:eu-west-2:123456789012:Firehose',
