+++ event-processing/lambda/tests/unit/app.event-processor.test.ts
@@ -6,6 +6,7 @@ import { AuditEvent as UnknownAuditEvent } from '../../tests/test-events/unknown
 import { SNS } from "aws-sdk";
 import {MockedFunction} from "ts-jest";
 import { randomUUID } from 'crypto';
+import {EventProcessorHelper} from "../test-helpers/event-processor-helper";
 
 jest.mock('aws-sdk', () => {
     const mockSNSInstance = {
@@ -109,35 +110,9 @@ describe('Unit test for app eventProcessorHandler', function () {
 
     it('successfully stringifies an SQS event', async () => {
         const expectedResult =
-            '{"event_id":"66258f3e-82fc-4f61-9ba0-62424e1f06b4","client_id":"some-client","timestamp":1609462861,"timestamp_formatted":"2021-01-23T15:43:21.842","event_name":"AUTHENTICATION_ATTEMPT","component_id":"1234","user":{"transaction_id":"a52f6f87","user_id":"some_user_id","email":"foo@bar.com","phone":"07711223344","ip_address":"100.100.100.100","session_id":"c222c1ec","persistent_session_id":"some session id","govuk_signin_journey_id":"43143-233Ds-2823-283-dj299j1"},"platform":{"xray_trace_id":"24727sda4192"},"restricted":{"experian_ref":"DSJJSEE29392"},"extensions":{"response":"Authentication successful"}}';
+            '{"event_id":"66258f3e-82fc-4f61-9ba0-62424e1f06b4","client_id":"some-client","timestamp":1609462861,"timestamp_formatted":"2021-01-23T15:43:21.842","event_name":"AUTHENTICATION_ATTEMPT","component_id":"AUTH","user":{"transaction_id":"a52f6f87","user_id":"some_user_id","email":"foo@bar.com","phone":"07711223344","ip_address":"100.100.100.100","session_id":"c222c1ec","persistent_session_id":"some session id","govuk_signin_journey_id":"43143-233Ds-2823-283-dj299j1"},"platform":{"xray_trace_id":"24727sda4192"},"restricted":{"experian_ref":"DSJJSEE29392"},"extensions":{"response":"Authentication successful"}}';
 
-        const exampleMessage: IAuditEvent = {
-            event_id: '66258f3e-82fc-4f61-9ba0-62424e1f06b4',
-            client_id: 'some-client',
-            timestamp: 1609462861,
-            timestamp_formatted: '2021-01-23T15:43:21.842',
-            event_name: 'AUTHENTICATION_ATTEMPT',
-            component_id: '1234',
-            user: {
-                transaction_id: 'a52f6f87',
-                user_id: 'some_user_id',
-                email: 'foo@bar.com',
-                phone: '07711223344',
-                ip_address: '100.100.100.100',
-                session_id: 'c222c1ec',
-                persistent_session_id: 'some session id',
-                govuk_signin_journey_id: '43143-233Ds-2823-283-dj299j1',
-            },
-            platform: {
-                xray_trace_id: '24727sda4192',
-            },
-            restricted: {
-                experian_ref: 'DSJJSEE29392',
-            },
-            extensions: {
-                response: 'Authentication successful',
-            },
-        };
+        const exampleMessage: IAuditEvent = EventProcessorHelper.exampleAuditMessage;
 
         (sns.publish().promise as MockedFunction<any>).mockResolvedValueOnce({Success: 'OK', MessageId: "1" });
 
@@ -164,35 +139,9 @@ describe('Unit test for app eventProcessorHandler', function () {
 
     it('successfully stringifies multiple events', async () => {
         const expectedResult =
-            '{"event_id":"66258f3e-82fc-4f61-9ba0-62424e1f06b4","client_id":"some-client","timestamp":1609462861,"timestamp_formatted":"2021-01-23T15:43:21.842","event_name":"AUTHENTICATION_ATTEMPT","component_id":"1234","user":{"transaction_id":"a52f6f87","user_id":"some_user_id","email":"foo@bar.com","phone":"07711223344","ip_address":"100.100.100.100","session_id":"c222c1ec","persistent_session_id":"some session id","govuk_signin_journey_id":"43143-233Ds-2823-283-dj299j1"},"platform":{"xray_trace_id":"24727sda4192"},"restricted":{"experian_ref":"DSJJSEE29392"},"extensions":{"response":"Authentication successful"}}';
+            '{"event_id":"66258f3e-82fc-4f61-9ba0-62424e1f06b4","client_id":"some-client","timestamp":1609462861,"timestamp_formatted":"2021-01-23T15:43:21.842","event_name":"AUTHENTICATION_ATTEMPT","component_id":"AUTH","user":{"transaction_id":"a52f6f87","user_id":"some_user_id","email":"foo@bar.com","phone":"07711223344","ip_address":"100.100.100.100","session_id":"c222c1ec","persistent_session_id":"some session id","govuk_signin_journey_id":"43143-233Ds-2823-283-dj299j1"},"platform":{"xray_trace_id":"24727sda4192"},"restricted":{"experian_ref":"DSJJSEE29392"},"extensions":{"response":"Authentication successful"}}';
 
-        const exampleMessage: IAuditEvent = {
-            event_id: '66258f3e-82fc-4f61-9ba0-62424e1f06b4',
-            client_id: 'some-client',
-            timestamp: 1609462861,
-            timestamp_formatted: '2021-01-23T15:43:21.842',
-            event_name: 'AUTHENTICATION_ATTEMPT',
-            component_id: '1234',
-            user: {
-                transaction_id: 'a52f6f87',
-                user_id: 'some_user_id',
-                email: 'foo@bar.com',
-                phone: '07711223344',
-                ip_address: '100.100.100.100',
-                session_id: 'c222c1ec',
-                persistent_session_id: 'some session id',
-                govuk_signin_journey_id: '43143-233Ds-2823-283-dj299j1',
-            },
-            platform: {
-                xray_trace_id: '24727sda4192',
-            },
-            restricted: {
-                experian_ref: 'DSJJSEE29392',
-            },
-            extensions: {
-                response: 'Authentication successful',
-            },
-        };
+        const exampleMessage: IAuditEvent = EventProcessorHelper.exampleAuditMessage;
 
         (sns.publish().promise as MockedFunction<any>).mockResolvedValueOnce({Success: 'OK', MessageId: "1" });
         (sns.publish().promise as MockedFunction<any>).mockResolvedValueOnce({Success: 'OK', MessageId: "2" });
@@ -222,7 +171,7 @@ describe('Unit test for app eventProcessorHandler', function () {
 
     it('successfully removes unrecognised elements from an audit event and user field and then logs to cloudwatch', async () => {
         const expectedResult =
-            '{"event_id":"66258f3e-82fc-4f61-9ba0-62424e1f06b4","client_id":"some-client","timestamp":1609462861,"timestamp_formatted":"2021-01-23T15:43:21.842","event_name":"AUTHENTICATION_ATTEMPT","component_id":"1234","user":{"transaction_id":"a52f6f87","user_id":"some_user_id","email":"foo@bar.com","phone":"07711223344","ip_address":"100.100.100.100","session_id":"c222c1ec","persistent_session_id":"some session id","govuk_signin_journey_id":"43143-233Ds-2823-283-dj299j1"},"platform":{"xray_trace_id":"24727sda4192"},"restricted":{"experian_ref":"DSJJSEE29392"},"extensions":{"response":"Authentication successful"}}';
+            '{"event_id":"66258f3e-82fc-4f61-9ba0-62424e1f06b4","client_id":"some-client","timestamp":1609462861,"timestamp_formatted":"2021-01-23T15:43:21.842","event_name":"AUTHENTICATION_ATTEMPT","component_id":"AUTH","user":{"transaction_id":"a52f6f87","user_id":"some_user_id","email":"foo@bar.com","phone":"07711223344","ip_address":"100.100.100.100","session_id":"c222c1ec","persistent_session_id":"some session id","govuk_signin_journey_id":"43143-233Ds-2823-283-dj299j1"},"platform":{"xray_trace_id":"24727sda4192"},"restricted":{"experian_ref":"DSJJSEE29392"},"extensions":{"response":"Authentication successful"}}';
 
         const exampleMessage: UnknownAuditEvent = {
             event_id: '66258f3e-82fc-4f61-9ba0-62424e1f06b4',
@@ -230,7 +179,7 @@ describe('Unit test for app eventProcessorHandler', function () {
             timestamp: 1609462861,
             timestamp_formatted: '2021-01-23T15:43:21.842',
             event_name: 'AUTHENTICATION_ATTEMPT',
-            component_id: '1234',
+            component_id: 'AUTH',
             user: {
                 transaction_id: 'a52f6f87',
                 user_id: 'some_user_id',
@@ -281,7 +230,7 @@ describe('Unit test for app eventProcessorHandler', function () {
 
     it('successfully populates missing formatted timestamp fields', async () => {
         const expectedResult =
-            '{"event_id":"66258f3e-82fc-4f61-9ba0-62424e1f06b4","client_id":"some-client","timestamp":1609462861,"timestamp_formatted":"2021-01-01T01:01:01.000Z","event_name":"AUTHENTICATION_ATTEMPT","component_id":"1234","user":{"transaction_id":"a52f6f87","user_id":"some_user_id","email":"foo@bar.com","phone":"07711223344","ip_address":"100.100.100.100","session_id":"c222c1ec","persistent_session_id":"some session id","govuk_signin_journey_id":"43143-233Ds-2823-283-dj299j1"},"platform":{"xray_trace_id":"24727sda4192"},"restricted":{"experian_ref":"DSJJSEE29392"},"extensions":{"response":"Authentication successful"}}';
+            '{"event_id":"66258f3e-82fc-4f61-9ba0-62424e1f06b4","client_id":"some-client","timestamp":1609462861,"timestamp_formatted":"2021-01-01T01:01:01.000Z","event_name":"AUTHENTICATION_ATTEMPT","component_id":"AUTH","user":{"transaction_id":"a52f6f87","user_id":"some_user_id","email":"foo@bar.com","phone":"07711223344","ip_address":"100.100.100.100","session_id":"c222c1ec","persistent_session_id":"some session id","govuk_signin_journey_id":"43143-233Ds-2823-283-dj299j1"},"platform":{"xray_trace_id":"24727sda4192"},"restricted":{"experian_ref":"DSJJSEE29392"},"extensions":{"response":"Authentication successful"}}';
 
         const exampleMessage: IAuditEvent = {
             event_id: '66258f3e-82fc-4f61-9ba0-62424e1f06b4',
@@ -289,7 +238,7 @@ describe('Unit test for app eventProcessorHandler', function () {
             timestamp: 1609462861,
             timestamp_formatted: '',
             event_name: 'AUTHENTICATION_ATTEMPT',
-            component_id: '1234',
+            component_id: 'AUTH',
             user: {
                 transaction_id: 'a52f6f87',
                 user_id: 'some_user_id',
@@ -336,35 +285,9 @@ describe('Unit test for app eventProcessorHandler', function () {
 
     it('logs an error when validation fails on event name', async () => {
         const expectedResult =
-            '{"event_id":"66258f3e-82fc-4f61-9ba0-62424e1f06b4","client_id":"some-client","timestamp":1609462861,"timestamp_formatted":"2021-01-23T15:43:21.842","event_name":"AUTHENTICATION_ATTEMPT","component_id":"1234","user":{"transaction_id":"a52f6f87","user_id":"some_user_id","email":"foo@bar.com","phone":"07711223344","ip_address":"100.100.100.100","session_id":"c222c1ec","persistent_session_id":"some session id","govuk_signin_journey_id":"43143-233Ds-2823-283-dj299j1"},"platform":{"xray_trace_id":"24727sda4192"},"restricted":{"experian_ref":"DSJJSEE29392"},"extensions":{"response":"Authentication successful"}}';
+            '{"event_id":"66258f3e-82fc-4f61-9ba0-62424e1f06b4","client_id":"some-client","timestamp":1609462861,"timestamp_formatted":"2021-01-23T15:43:21.842","event_name":"AUTHENTICATION_ATTEMPT","component_id":"AUTH","user":{"transaction_id":"a52f6f87","user_id":"some_user_id","email":"foo@bar.com","phone":"07711223344","ip_address":"100.100.100.100","session_id":"c222c1ec","persistent_session_id":"some session id","govuk_signin_journey_id":"43143-233Ds-2823-283-dj299j1"},"platform":{"xray_trace_id":"24727sda4192"},"restricted":{"experian_ref":"DSJJSEE29392"},"extensions":{"response":"Authentication successful"}}';
 
-        const exampleMessage: IAuditEvent = {
-            event_id: '66258f3e-82fc-4f61-9ba0-62424e1f06b4',
-            client_id: 'some-client',
-            timestamp: 1609462861,
-            timestamp_formatted: '2021-01-23T15:43:21.842',
-            event_name: 'AUTHENTICATION_ATTEMPT',
-            component_id: '1234',
-            user: {
-                transaction_id: 'a52f6f87',
-                user_id: 'some_user_id',
-                email: 'foo@bar.com',
-                phone: '07711223344',
-                ip_address: '100.100.100.100',
-                session_id: 'c222c1ec',
-                persistent_session_id: 'some session id',
-                govuk_signin_journey_id: '43143-233Ds-2823-283-dj299j1',
-            },
-            platform: {
-                xray_trace_id: '24727sda4192',
-            },
-            restricted: {
-                experian_ref: 'DSJJSEE29392',
-            },
-            extensions: {
-                response: 'Authentication successful',
-            },
-        };
+        const exampleMessage: IAuditEvent = EventProcessorHelper.exampleAuditMessage;
 
         const exampleInvalidMessage: IAuditEvent = {
             event_id: '66258f3e-82fc-4f61-9ba0-62424e1f06b4',
@@ -427,35 +350,9 @@ describe('Unit test for app eventProcessorHandler', function () {
 
     it('logs an error when validation fails on timestamp', async () => {
         const expectedResult =
-            '{"event_id":"66258f3e-82fc-4f61-9ba0-62424e1f06b4","client_id":"some-client","timestamp":1609462861,"timestamp_formatted":"2021-01-23T15:43:21.842","event_name":"AUTHENTICATION_ATTEMPT","component_id":"1234","user":{"transaction_id":"a52f6f87","user_id":"some_user_id","email":"foo@bar.com","phone":"07711223344","ip_address":"100.100.100.100","session_id":"c222c1ec","persistent_session_id":"some session id","govuk_signin_journey_id":"43143-233Ds-2823-283-dj299j1"},"platform":{"xray_trace_id":"24727sda4192"},"restricted":{"experian_ref":"DSJJSEE29392"},"extensions":{"response":"Authentication successful"}}';
+            '{"event_id":"66258f3e-82fc-4f61-9ba0-62424e1f06b4","client_id":"some-client","timestamp":1609462861,"timestamp_formatted":"2021-01-23T15:43:21.842","event_name":"AUTHENTICATION_ATTEMPT","component_id":"AUTH","user":{"transaction_id":"a52f6f87","user_id":"some_user_id","email":"foo@bar.com","phone":"07711223344","ip_address":"100.100.100.100","session_id":"c222c1ec","persistent_session_id":"some session id","govuk_signin_journey_id":"43143-233Ds-2823-283-dj299j1"},"platform":{"xray_trace_id":"24727sda4192"},"restricted":{"experian_ref":"DSJJSEE29392"},"extensions":{"response":"Authentication successful"}}';
 
-        const exampleMessage: IAuditEvent = {
-            event_id: '66258f3e-82fc-4f61-9ba0-62424e1f06b4',
-            client_id: 'some-client',
-            timestamp: 1609462861,
-            timestamp_formatted: '2021-01-23T15:43:21.842',
-            event_name: 'AUTHENTICATION_ATTEMPT',
-            component_id: '1234',
-            user: {
-                transaction_id: 'a52f6f87',
-                user_id: 'some_user_id',
-                email: 'foo@bar.com',
-                phone: '07711223344',
-                ip_address: '100.100.100.100',
-                session_id: 'c222c1ec',
-                persistent_session_id: 'some session id',
-                govuk_signin_journey_id: '43143-233Ds-2823-283-dj299j1',
-            },
-            platform: {
-                xray_trace_id: '24727sda4192',
-            },
-            restricted: {
-                experian_ref: 'DSJJSEE29392',
-            },
-            extensions: {
-                response: 'Authentication successful',
-            },
-        };
+        const exampleMessage: IAuditEvent = EventProcessorHelper.exampleAuditMessage;
 
         const exampleInvalidMessage: IAuditEvent = {
             event_id: '66258f3e-82fc-4f61-9ba0-62424e1f06b4',
